优化像素地形生成逻辑

+++ Source/WHFramework/Private/Common/CommonStatics.cpp
@@ -215,7 +215,7 @@ bool UCommonStatics::StringToBool(const FString& InString)
 	return FCoreStatics::StringToBool(InString);
 }
 
-FString UCommonStatics::SanitizeFloat(float InFloat, int32 InMaxDigits)
+FString UCommonStatics::SanitizeFloat(double InFloat, int32 InMaxDigits)
 {
 	return FCoreStatics::SanitizeFloat(InFloat, InMaxDigits);
 }
