+++ Source/WHFrameworkCore/Public/WHFrameworkCoreStatics.h
@@ -81,7 +81,7 @@ public:
 
 	static bool StringToBool(const FString& InString);
 	
-	static FString SanitizeFloat(float InFloat, int32 InMaxDigits = -1);
+	static FString SanitizeFloat(double InFloat, int32 InMaxDigits = -1);
 
 	//////////////////////////////////////////////////////////////////////////
 	// Url
